Creates a description file for a compendium

The idea behind a compendium is to have a minimal description file that makes it easy for anyone to 'install' your analysis dependencies. This makes it possible for someone to run your code easily.

write_compendium_description(type = "Compendium",
  package = "Compendium title", description = "Compendium description",
  version = "0.0.1", path = ".")

Arguments

type
Default here is compendium

package
Name of your compendium

description
Description of your compendium

version
Version of your compendium

path
path to project (in case it is not in the current working directory)

Details

To automatically populate author information, you may set usethis options in your `.rprofile` like so. options( usethis.full_name = "Karthik Ram", usethis.description = list( `Authors@R` = 'person("Karthik", "Ram", email = "karthik.ram@gmail.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-0233-1757"))', License = "MIT + file LICENSE", Version = "0.0.0.9000" ) )