Write a minimal Dockerfile for a Binder

Write a minimal Dockerfile for a Binder

write_dockerfile(base = NULL, path = ".",
  maintainer = getOption("usethis.full_name"), r_date = NULL)

Arguments

base
Your Docker base. I recommend that you use a binder base from the rocker project. This daily image will contain all the Jupyter hub elements + Rstudio Server + the tidyverse (which cuts down on installation time). The only thing then left to do is to install any additional packages listed in your DESCRIPTION file, which will be done during the build_binder step.

path
path to binder repo. Defaults to current location.

maintainer
Maintainer of the Dockerfile. By default the function reads `usethis.full_name` set in Options. The same information is used to set the author on your DESCRIPTION file. For more information on setting up default values for Description files, reasd the Rstudio usethis documentation.

r_date
The date for which you'd like to lock down this project. Projects that match current release date go to "latest". This will be automatically determined from your code, so the recommendation is not to override this unless you have a specific reason to do so. Date must be in ISO 8601 format.